Update to most recent changes
I intend for this to be the last of my series of disorganized commits on
this repo. From now on, I hope to be able to maintain precise tracking
of each resource separately, maybe even trying out git-flow (I don't
deem it necessary at all for a repo like this, but it will make nice
practice).